Update ncurses to fix input issues on tmux 3.3a
Updating to 3.3a[1], with the default option set[2], tmux started
producing strange behavior. The symptoms included not registering
backspaces, and using the `<UP>` key appended the previous command to
the current one in the shell.

A [discussion on the Homebrew
repository](https://github.com/orgs/Homebrew/discussions/3368) produced
some suggestions. Although updating zsh didn't work, setting
`default-terminal` from `tmux-256color` to `screen-256color` resolved
the issue:

    set -g default-terminal screen-256color

However, another comment suggested that the change to `tmux-265color` as
the `default-terminal` was a recent change to tmux, and that macOS ships
with a very old version of ncurses. Updating ncurses resolved the issue,
even when removing the `default-terminal` configuration option again.

[1]:
    tmux -V

    tmux 3.3a

[2]:
    tmux show-options -s

    backspace C-?
    buffer-limit 50
    command-alias[0] split-pane=split-window
    command-alias[1] splitp=split-window
    command-alias[2] "server-info=show-messages -JT"
    command-alias[3] "info=show-messages -JT"
    command-alias[4] "choose-window=choose-tree -w"
    command-alias[5] "choose-session=choose-tree -s"
    copy-command ''
    default-terminal tmux-256color
    editor nvim
    escape-time 500
    exit-empty on
    exit-unattached off
    extended-keys off
    focus-events off
    history-file ''
    message-limit 1000
    prompt-history-limit 100
    set-clipboard external
    terminal-overrides
    terminal-features[0] xterm*:clipboard:ccolour:cstyle:focus:title
    terminal-features[1] screen*:title
    user-keys