Set tmux escape-time to 0
By default, tmux uses a 500 millisecond `escape-time`, which causes
problems with multi-function keyboard keys[1]. Set it to 0 instead:

    set  -s escape-time 0

[1] For example, I have my `<caps lock>` key mapped to `<ctrl>` on hold,
and `<escape>` on tap. With tmux's default escape time, tapping it
without waiting for half a second before the next key press results in a
`<ctrl>` being sent instead of an `<escape>`.